+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.20] - 2026-01-22
+
+### Added
+- **`adelete_thread()` and `delete_thread()`** methods in `SimpleRedisCheckpointer`
+  - Delete all checkpoints for a given thread ID
+  - Uses SCAN to find all keys matching thread pattern
+  - Required for session cleanup in applications using Redis checkpointer
+  - 3 new unit tests added
+
 ## [0.3.19] - 2026-01-22
 
 ### Added